changed ints to long long for reading large trajectory files

+++ analysis/src/oxDNA_analysis_tools/cython_utils/get_confs.pyx
@@ -31,11 +31,11 @@ def cget_confs(list idxs, str traj_path, int start, int nconfs, int nbases, bint
 
     # Configuration start/size markers within the chunk
     cdef int *sizes = <int *> malloc(cnconfs * sizeof(int))
-    cdef int *conf_starts = <int *> malloc(cnconfs * sizeof(int))
+    cdef long long *conf_starts = <long long *> malloc(cnconfs * sizeof(long long))
     if not sizes or not conf_starts:
         raise MemoryError("Could not allocate memory for the configuration sizes and starts")
 
-    cdef int chunk_size = idxs[start+cnconfs-1].offset + idxs[start+cnconfs-1].size - idxs[start].offset
+    cdef long long chunk_size = idxs[start+cnconfs-1].offset + idxs[start+cnconfs-1].size - idxs[start].offset
     for i in range(cnconfs):
         sizes[i] = idxs[start+i].size
         conf_starts[i] = idxs[start+i].offset - idxs[start].offset
